- Use latest open graphy library version to fix rendering via web server on ubuntu arm systems. - Improve the deletion process of products that can be deleted (never ordered).

+++ app/Models/OneTimeProduct.php
@@ -44,15 +44,31 @@ class OneTimeProduct extends Model
         });
 
         static::deleting(function (OneTimeProduct $oneTimeProduct) {
+            if (!$oneTimeProduct->isDeletable()) {
+                throw new \Exception('Cannot delete a one-time product that has been ordered.');
+            }
+
             $oneTimeProduct->paymentProviderData()->delete();
             foreach ($oneTimeProduct->prices as $price) {
                 $price->pricePaymentProviderData()->delete();
             }
+
+            $oneTimeProduct->prices()->delete();
         });
+    }
+
+    public function isDeletable()
+    {
+        return !$this->orderItems()->exists();
     }
 
     public function paymentProviderData()
     {
         return $this->hasMany(OneTimeProductPaymentProviderData::class);
+    }
+
+    public function orderItems()
+    {
+        return $this->hasMany(OrderItem::class);
     }
 }
